[cheri][loader] Add support for CHERI
This commit adds support to load CHERI purecap ELFs and protection
domains. The loader itself isn't purecap and only needs to be built
in hybrid mode so that it can create valid CHERI capability pointers
for the monitor's task, whether it's purecap or not. It also gets the
size of the code segement so that it applies proper bounds in case it
is built in purecap mode.

Signed-off-by: Hesham Almatary <heshamalmatary@capabilitieslimited.co.uk>