[cheri][microkit] Enable RISC-V PTE reads/writes
CHERI extends RISC-V's PTE permissions to enable/disable CHERI cap
writes (and implicitly reads) per-page. For CHERI-seL4 and
CHERI-Microkit, we globablly allow that by default for everything as
there's no current use-case when this needs to be disabled.

CHERI also adds a CRG bit to PTEs in order to implement temporal-safety
techniques (eg revocations). We also don't support that at the moment.

Signed-off-by: Hesham Almatary <heshamalmatary@capabilitieslimited.co.uk>