libutils: Add memcpy-family and printf functions
Copy over memzero and memset from the seL4 kernel. memcpy has to be CHERI-aware,
so it is copied from CHERI-newlib.

printf is the same from the Microkit's VMM example, it is copied as is but
also add support for printing formatted CHERI capabilities, this code is
copied from Arm's Morello muslc and is modified to work for standard
CHERI-RISC-V.

Signed-off-by: Hesham Almatary <heshamalmatary@capabilitieslimited.co.uk>