[cheri][microkit] Account for multi-page ELF MRs
Page descriptions are repeated for multi-page ELF MR regions. Thus, if a symbol
exists in a multipage ELF MR, its page won't be mapped to an exactly one page
with the provided page-aligned vaddr. Instead, there will be just one entry
that holds the first page details of that MR, and repeated for x counts times,
depending on the region size. This commit accounts for this case, in situations
where symbols are placed in such regions.

Signed-off-by: Hesham Almatary <heshamalmatary@capabilitieslimited.co.uk>