[cheri][microkit] Make root codecap cover captable
MKpuABI handes over a root codecap to new threads to construct captables from.
As this is still PCC-relative, loading from the captable is also PCC-relative in
which case the captable needs to be within PCC's bounds.

This commit adds the data segment, in addition to the code one, to the root
codecap bounds, similar to how PCC is constructed.

Signed-off-by: Hesham Almatary <heshamalmatary@capabilitieslimited.co.uk>